por Raphael Menezes: Adaptando o botao de remover arquivos no show de curso

+++ db/migrate/20140120181417_create_fases.rb
@@ -3,7 +3,7 @@ class CreateFases < ActiveRecord::Migration
     create_table :fases do |t|
       t.string :descricao
       t.integer :curso_id, :null => false, :references => [:cursos, :id]
-      t.float :orcamento
+      t.integer :orcamento
       t.timestamps
     end
   end
